remove response postprocessing & fixes
