--- src/index.ts
@@ -1,1 +1,0 @@
-export { JSP } from './classes/JSP.ts';
