+++ src/types/Request.ts
@@ -1,0 +1,6 @@
+export type RequestRequestOptions = {
+	endpoint: string;
+	method: 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE';
+	headers?: Record<string, string>;
+	body?: string;
+};
