--- src/types/request/document/AccessOptions.ts
@@ -1,5 +1,0 @@
-type AccessOptions = {
-	password?: string;
-};
-
-export type { AccessOptions };
