--- src/types/request/document/EditOptions.ts
@@ -1,7 +1,0 @@
-type EditOptions = {
-	newBody: string;
-	password?: string;
-	secret: string;
-};
-
-export type { EditOptions };
