--- src/types/request/document/PublishOptions.ts
@@ -1,9 +1,0 @@
-type PublishOptions = {
-	key?: string;
-	keyLength?: string;
-	secret?: string;
-	password?: string;
-	lifetime?: string;
-};
-
-export type { PublishOptions };
