--- src/types/request/document/RemoveOptions.ts
@@ -1,5 +1,0 @@
-type RemoveOptions = {
-	secret: string;
-};
-
-export type { RemoveOptions };
