--- src/types/response/AccessedDocument.ts
@@ -1,8 +1,0 @@
-type AccessedDocument = {
-	key: string;
-	data: string;
-	url?: string;
-	expirationTimestamp?: number;
-};
-
-export type { AccessedDocument };
