--- src/types/response/JSPError.ts
@@ -1,7 +1,0 @@
-type JSPError = {
-	type: string;
-	errorCode: string;
-	message: string;
-};
-
-export type { JSPError };
