--- src/types/response/PublishedDocument.ts
@@ -1,8 +1,0 @@
-type PublishedDocument = {
-	key: string;
-	secret: string;
-	url?: string;
-	expirationTimestamp?: number;
-};
-
-export type { PublishedDocument };
