+++ tsconfig.json
@@ -4,7 +4,7 @@
 		"lib": ["ESNext"],
 		"module": "ESNext",
 		"moduleResolution": "Bundler",
-		"target": "ESNext",
+		"target": "ES2021",
 
 		"allowImportingTsExtensions": true,
 		"allowJs": true,
